+++ detprocess/core/__init__.py
@@ -1,0 +1,3 @@
+from .features import *
+from .oftrigger import *
+from .eventbuilder import *
