+++ detprocess/core/oftrigger.py
