+++ detprocess/process/__init__.py
@@ -1,7 +1,4 @@
-from ._process import *
-from .process_trigger import *
-from ._processing_data import *
-from ._features import *
+from .process_features import *
+from .process_triggers import *
+from .processing_data import *
 from .randoms import *
-from .oftrigger import *
-from .eventbuilder import *
